+++ doc/conf.py
@@ -29,9 +29,9 @@ copyright = '2018-2020, SISSA (International School for Advanced Studies)'
 author = 'Davide Brunato'
 
 # The short X.Y version
-version = '2.0'
+version = '2.1'
 # The full version, including alpha/beta/rc tags
-release = '2.0.5'
+release = '2.1.0'
 
 
 # -- General configuration ---------------------------------------------------